Crate hopr_protocol_hopr

Modules

prelude

Structs

AuxiliaryPacketInfo

Contains some miscellaneous information about a received packet.

FoundSurb

Contains a SURB found in the SURB ring buffer via [SurbStore::find_surb].

HoprCodecConfig

Configuration of HoprEncoder and HoprDecoder.

HoprDecoder

Default decoder implementation for HOPR packets.

HoprEncoder

Default encoder implementation for HOPR packets.

HoprTicketProcessor

HOPR-specific implementation of UnacknowledgedTicketProcessor and TicketTracker.

HoprTicketProcessorConfig

Configuration for the HOPR ticket processor within the packet pipeline.

IncomingAcknowledgementPacket

Incoming packet that contains acknowledgements of delivered packets.

IncomingFinalPacket

An incoming packet with a payload intended for us.

IncomingForwardedPacket

Incoming packet that must be forwarded.

MemorySurbStore

Basic SurbStore implementation based on an in-memory cache.

OutgoingPacket

Packet that is being sent out by us.

SurbStoreConfig

Configuration for the SURB cache.

Enums

HoprProtocolError

IncomingPacket

Incoming HOPR packet.

IncomingPacketError

Error that can occur when processing an incoming packet.

ResolvedAcknowledgement

Determines the result of how an acknowledgement was resolved.

TicketAcknowledgementError

Defines errors returned by [UnacknowledgedTicketProcessor::acknowledge_ticket].

TicketCreationError

Error that can occur when creating a ticket.

Constants

MAX_ACKNOWLEDGEMENTS_BATCH_SIZE

Maximum number of acknowledgements that can be packed into a single HOPR packet.

Traits

PacketDecoder

Trait defining decoder HOPR packets.

PacketEncoder

Trait defining encoder for outgoing HOPR packets.

SurbStore

A trait defining the operations required to store and retrieve SURBs (Single Use Reply Blocks) and their reply openers.

TicketTracker

Allows tracking ticket indices of outgoing channels and unrealized balances of incoming channels.

UnacknowledgedTicketProcessor

Performs necessary processing of unacknowledged tickets in the HOPR packet processing pipeline.