hopr_protocol_hopr

Module traits

Re-exports

pub use crate::errors::IncomingPacketError;

pub use crate::types::FoundSurb;

pub use crate::types::IncomingPacket;

pub use crate::types::OutgoingPacket;

pub use crate::types::ResolvedAcknowledgement;

Enums

TicketAcknowledgementError

Defines errors returned by UnacknowledgedTicketProcessor::acknowledge_ticket.

Traits

PacketDecoder

Trait defining decoder HOPR packets.

PacketEncoder

Trait defining encoder for outgoing HOPR packets.

SurbStore

A trait defining the operations required to store and retrieve SURBs (Single Use Reply Blocks) and their reply openers.

UnacknowledgedTicketProcessor

Performs necessary processing of unacknowledged tickets in the HOPR packet processing pipeline.