hopr_protocol_hopr

Module types

Structs

AuxiliaryPacketInfo

Contains some miscellaneous information about a received packet.

FoundSurb

Contains a SURB found in the SURB ring buffer via SurbStore::find_surb.

IncomingAcknowledgementPacket

Incoming packet that contains acknowledgements of delivered packets.

IncomingFinalPacket

An incoming packet with a payload intended for us.

IncomingForwardedPacket

Incoming packet that must be forwarded.

OutgoingPacket

Packet that is being sent out by us.

Enums

IncomingPacket

Incoming HOPR packet.

ResolvedAcknowledgement

Determines the result of how an acknowledgement was resolved.